Add previously_migrated_as? check to renamed migrations

+++ db/migrate/20160428215838_remove_filters_from_miq_groups.rb
@@ -1,5 +1,8 @@
 class RemoveFiltersFromMiqGroups < ActiveRecord::Migration[5.0]
+  include MigrationHelper
+
   def change
+    return if previously_migrated_as?("20160414132130")
     remove_column :miq_groups, :filters, :text
   end
 end
